experiments can now be executed in two modes: grid and list.
add a parameter experiment = 'grid' or 'list' to the experiment section to switch between the two modes. default is 'grid' mode, like before.
this applies to parameter lists only. in grid mode, every single combination of parameters is executed. in list mode, the parameter lists are zipped together (the total number is the length of the shortest list). See experiments.cfg for examples.